requirements.txt
Flask
gspread
oauth2client
qrcode
python-dotenv
gunicorn
